fixing printx

+++ ft_printf.c
@@ -247,7 +247,7 @@ void printNums(t_env *env, va_list ap, int base, int upperCase)
 	char *str;
 
 	d = va_arg(ap, void *);
-	str = ft_itoa_base((intmax_t)d, base, upperCase);
+	str = ft_itoa_base((long long)d, base, upperCase);
 	if (env->plus && d >= 0)
 	{
 		write(1, "+", 1);
